add user and enrich activator with user data, and some refactoring work
